fix: Reduce the write lock acquisition time
This commit should fix #113 where the `ClearLocalMessageHistoryAction.update` was
causing a `PluginException: 316 ms to call on EDT`
because the `InternalVcsService` was slow to get created.

The reason is that the creation happened on EDT, but the thread couldn't progress because there was a lock.
The idea here is to reduce the contention by moving the heavy computation outside the write lock.

Refs: #113